# Conflicts:
#	app/src/main/java/one/mixin/android/ui/web/WebFragment.kt

+++ app/src/main/java/one/mixin/android/job/MessageDeleteJob.kt
@@ -3,44 +3,46 @@ package one.mixin.android.job
 import com.birbit.android.jobqueue.Params
 import kotlinx.coroutines.runBlocking
 import one.mixin.android.Constants.DB_DELETE_LIMIT
-import one.mixin.android.db.deleteMessageByConversationId
+import one.mixin.android.db.deleteMessageByIds
+import one.mixin.android.db.flow.MessageFlow
+import one.mixin.android.fts.deleteByMessageIds
 
 class MessageDeleteJob(
     private val conversationId: String,
-    private val deleteMention: Boolean = false,
-    private val deleteConversation: Boolean = false
+    private val lastRowId: Long,
+    private val deleteConversation: Boolean,
 ) :
     BaseJob(Params(PRIORITY_UI_HIGH).addTags(GROUP).groupBy("message_delete").persist()) {
-
     private val TAG = MessageDeleteJob::class.java.simpleName
 
     companion object {
         const val GROUP = "MessageDeleteJob"
-        private const val serialVersionUID = 1L
+        private const val serialVersionUID = 2L
     }
 
-    override fun onRun() = runBlocking {
-        if (deleteMention) {
-            val deleteTimes =
-                messageMentionDao.countDeleteMessageByConversationId(conversationId) / DB_DELETE_LIMIT + 1
-            repeat(deleteTimes) {
-                messageMentionDao.deleteMessageByConversationId(conversationId, DB_DELETE_LIMIT)
-            }
-        } else {
+    override fun onRun() =
+        runBlocking {
             val deleteTimes =
                 messageDao.countDeleteMessageByConversationId(conversationId) / DB_DELETE_LIMIT + 1
             repeat(deleteTimes) {
-                messageFts4Dao.deleteByMessageIds(
+                val ids =
                     messageDao.getMessageIdsByConversationId(
                         conversationId,
-                        DB_DELETE_LIMIT
+                        lastRowId,
+                        DB_DELETE_LIMIT,
                     )
-                )
-                appDatabase.deleteMessageByConversationId(conversationId, DB_DELETE_LIMIT)
+                ftsDatabase.deleteByMessageIds(ids)
+                appDatabase.deleteMessageByIds(ids)
+                MessageFlow.delete(conversationId, ids)
+            }
+            val currentRowId = messageDao.findLastMessageRowId(conversationId)
+            if (deleteConversation && currentRowId == null) {
+                conversationDao.deleteConversationById(conversationId)
+                conversationExtDao.deleteConversationById(conversationId)
+            } else {
+                remoteMessageStatusDao.countUnread(conversationId)
+                conversationDao.refreshLastMessageId(conversationId)
+                conversationExtDao.refreshCountByConversationId(conversationId)
             }
         }
-        if (deleteConversation) {
-            conversationDao.deleteConversationById(conversationId)
-        }
-    }
 }
